requirements.txt
fastavro==1.12.1
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
pandavro==1.9.0
python-dotenv==1.2.1
scikit_learn==1.7.2
tqdm==4.67.1
